request manager will handle http requests move doRequests from webdav to request manager create bm (business manager) to separate from webdav
