+++ index.js
@@ -9,6 +9,7 @@ process.env.UV_THREADPOOL_SIZE = 128;
     Core: require('./src/davos'),
     ConfigManager: require('./src/config-manager'),
     Logger: require('./src/logger'),
+    BM: require('./src/bm'),
     WebDav: require('./src/webdav')
   };
 }());
